# schema/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchCreateRequest": {
        "description": "Create a batch scrape: multiple counties x record types under one parent,\nsharing fields/enrichment/deliver. Fans out into N child scrapes. The batch\nowns delivery (one combined CSV); per-child delivery is suppressed.\n\n2B: `schedule` (same validated shape as single scrapers) makes the batch\nrecur — the scheduler beat creates a new run when the schedule fires.\nDefault (frequency='manual') = 2A behavior: one immediate run, no recurrence.\nChildren always store schedule={} — the PARENT owns scheduling, and\ndispatch_scheduled_jobs skips frequency='manual' children.",
        "properties": {
          "counties": {
            "items": {
              "type": "string"
            },
            "maxItems": 250,
            "minItems": 1,
            "title": "Counties",
            "type": "array"
          },
          "deliver": {
            "$ref": "#/components/schemas/DeliverConfig",
            "default": {
              "emails": [],
              "formats": [
                "csv"
              ]
            }
          },
          "enrichment": {
            "$ref": "#/components/schemas/EnrichmentConfig",
            "default": {
              "property_lookup": true,
              "skip_tracing": false
            }
          },
          "fields": {
            "$ref": "#/components/schemas/FieldsConfig",
            "default": {
              "date_recorded": true,
              "heirs": true,
              "legal_description": false,
              "mailing_address": true,
              "parcel_id": true,
              "party_name": true,
              "property_address": true
            }
          },
          "include_living_owner_tod": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Living Owner Tod"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "record_types": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Record Types",
            "type": "array"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleConfig",
            "default": {
              "date_range_mode": "rolling_90",
              "frequency": "manual",
              "run_at_day_of_month": 1,
              "run_at_hour": 6,
              "run_at_minute": 0,
              "run_at_weekday": 0
            }
          },
          "skip_trace_enabled": {
            "default": false,
            "title": "Skip Trace Enabled",
            "type": "boolean"
          },
          "state": {
            "maxLength": 16,
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "counties",
          "record_types"
        ],
        "title": "BatchCreateRequest",
        "type": "object"
      },
...
      "BatchSummaryResponse": {
        "description": "A batch + its (single, on-demand 2A) run status — for the list view.",
        "properties": {
          "child_count": {
            "default": 0,
            "title": "Child Count",
            "type": "integer"
          },
          "combined_export_ready": {
            "default": false,
            "title": "Combined Export Ready",
            "type": "boolean"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "run_status": {
            "$ref": "#/components/schemas/BatchRunStatus",
            "default": "pending"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "created_at"
        ],
        "title": "BatchSummaryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "BridgeLeads API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/batches": {
      "get": {
        "description": "List the current user's batches, newest first, with run status.",
        "operationId": "list_batches_batches_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BatchSummaryResponse"
                  },
                  "title": "Response List Batches Batches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Batches",
        "tags": [
          "batches"
        ]
      },
      "post": {
        "operationId": "create_batch_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Batch",
        "tags": [
          "batches"
        ]
      }
    },
    "/batches/{batch_id}": {
      "get": {
        "description": "A batch's status + per-child (county x record_type) summary.",
        "operationId": "get_batch_batches__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Batch",
        "tags": [
          "batches"
        ]
      }
    },
    "/batches/{batch_id}/download": {
      "get": {
        "description": "Stream the batch's combined CSV through this authed endpoint.\n\nThe CSV is REBUILT from the DB on demand (not fetched from R2): the api has\nno R2 credentials (those live on the worker), and rebuilding means a\nre-download reflects later async skip-trace fills. Tenant-scoped by the\nverified owner's id. 404 until the barrier has finished (combined_export_key\nis the 'ready' marker).",
        "operationId": "download_batch_batches__batch_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Batch",
        "tags": [
          "batches"
        ]
      }
    },
    "/batches/{batch_id}/runs": {
      "get": {
        "description": "Run history for a batch, newest first (2B: runs are plural). The combined\nCSV per run is fetched via the run-scoped download below — never the R2 key.",
        "operationId": "list_batch_runs_batches__batch_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BatchRunResponse"
                  },
                  "title": "Response List Batch Runs Batches  Batch Id  Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Batch Runs",
        "tags": [
          "batches"
        ]
      }
    },
    "/batches/{batch_id}/runs/{run_id}/download": {
      "get": {
        "description": "Run-scoped combined-CSV download (2B): history downloads must not drift to\nthe latest run the way /download (latest-run semantics) does.",
        "operationId": "download_batch_run_batches__batch_id__runs__run_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Batch Run",
        "tags": [
          "batches"
        ]
      }
    },
...
  }
}